refactor(data): addendum
