+++ app/src/main/java/com/blueprint/featureA/mapper/DefaultFeatureALocalMapper.kt
@@ -1,7 +1,7 @@
 package com.blueprint.featureA.mapper
 
 import com.blueprint.featureA.data.local_model.LocalFeatureA
-import com.blueprint.featureA.domain.FeatureA
+import com.blueprint.featureA.domain.model.FeatureA
 
 class DefaultFeatureALocalMapper: FeatureALocalMapper {
     override fun map(featuresA: List<LocalFeatureA>): List<FeatureA> =
