+++ app/src/main/java/com/blueprint/featureA/mapper/DefaultFeatureARemoteRemoteMapper.kt
@@ -1,7 +1,7 @@
 package com.blueprint.featureA.mapper
 
 import com.blueprint.featureA.data.remote_model.RemoteFeatureA
-import com.blueprint.featureA.domain.FeatureA
+import com.blueprint.featureA.domain.model.FeatureA
 
 class DefaultFeatureARemoteRemoteMapper : FeatureARemoteMapper {
     override fun map(featuresA: List<RemoteFeatureA>): List<FeatureA> =
