+++ app/src/main/java/com/blueprint/featureA/repository/DefaultFeatureARemoteRepository.kt
@@ -1,7 +1,7 @@
 package com.blueprint.featureA.repository
 
 import com.blueprint.featureA.service.FeatureAService
-import com.blueprint.featureA.domain.FeatureA
+import com.blueprint.featureA.domain.model.FeatureA
 import com.blueprint.featureA.mapper.FeatureARemoteMapper
 import javax.inject.Inject
